django/requirements.txt
asgiref==3.4.1
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==35.0.0
cssselect2==0.5.0
cycler==0.11.0
dj-rest-auth==2.1.11
Django==4.0.4
django-colorfield==0.4.1
django-cors-headers==3.8.0
django-extensions==3.1.5
django-filter==21.1
django-js-asset==1.2.2
django-mptt==0.13.4
django-renderpdf==3.0.1
django-rest-passwordreset==1.2.1
django-simple-history==3.0.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
fonttools==4.29.1
graphviz==0.19.1
greenlet==1.1.2
html5lib==1.1
idna==3.3
kiwisolver==1.4.1
Markdown==3.3.4
matplotlib==3.5.1
numpy==1.23.0
packaging==21.3
pandas==1.4.3
Pillow==8.2.0
# psycopg2==2.9.3
# psycopg2-binary==2.9.3
pycparser==2.21
pydot==1.4.2
pydyf==0.1.2
PyJWT==1.7.1
pyparsing==2.4.7
pyphen==0.12.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
tablib==3.2.0
tinycss2==1.1.1
typing-extensions==3.10.0.2
urllib3==1.26.9
weasyprint==54.2
webencodings==0.5.1
zopfli==0.2.1
sqlalchemy
python-decouple
